requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
faiss-cpu
python-dotenv
openai
aiosqlite
duckduckgo-search
